refactor(component-linter): Consolidate overlapping rules and remove deprecated (63 → 55)
- Delete 2 deprecated empty rules: dependency-prop-validation,
  validate-component-props (both returned empty arrays)
- Delete 2 superseded rules: entity-field-validation (replaced by
  entity-field-access-validation), query-field-validation (replaced by
  query-result-field-access-validation)
- Consolidate 4 callback/event rules into callback-event-validation:
  callback-parameter-validation, callbacks-passthrough-only,
  callbacks-usage-validation, event-invocation-pattern
- Consolidate 2 styles rules into styles-validation:
  styles-invalid-path, styles-unsafe-access
