+++ packages/React/test-harness/src/lib/runtime-rules/index.ts
@@ -1,21 +1,16 @@
 // Auto-import all rule files to trigger @RegisterClass decorators
 import './ai-tools-availability-check';
-import './callback-parameter-validation';
-import './callbacks-passthrough-only';
-import './callbacks-usage-validation';
+import './callback-event-validation';
 import './chart-field-validation';
 import './child-component-prop-validation';
 import './component-name-mismatch';
 import './component-not-in-dependencies';
 import './component-props-validation';
 import './component-usage-without-destructuring';
-import './datagrid-field-validation';
 import './data-result-validation';
-import './dependency-prop-validation';
+import './datagrid-field-validation';
 import './dependency-shadowing';
 import './entity-field-access-validation';
-import './entity-field-validation';
-import './event-invocation-pattern';
 import './library-variable-names';
 import './no-child-implementation';
 import './no-data-prop';
@@ -33,7 +28,6 @@ import './prefer-async-await';
 import './prefer-jsx-syntax';
 import './prop-state-sync';
 import './property-name-consistency';
-import './query-field-validation';
 import './query-result-field-access-validation';
 import './react-component-naming';
 import './react-hooks-rules';
@@ -45,8 +39,7 @@ import './server-reload-on-client-operation';
 import './single-function-only';
 import './string-replace-all-occurrences';
 import './string-template-validation';
-import './styles-invalid-path';
-import './styles-unsafe-access';
+import './styles-validation';
 import './type-inference-errors';
 import './type-mismatch-operation';
 import './undefined-component-usage';
@@ -60,5 +53,4 @@ import './use-unwrap-components';
 import './useeffect-unstable-dependencies';
 import './utilities-api-validation';
 import './utilities-no-direct-instantiation';
-import './validate-component-props';
 import './validate-component-references';
